pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oracle"
version = "0.1.0"
description = "OTEL-native runtime, adapters, and materializers for evidence-first DSA workflows"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
otel = [
  "opentelemetry-api>=1.20.0",
  "opentelemetry-sdk>=1.20.0",
  "opentelemetry-exporter-otlp>=1.20.0",
]

[dependency-groups]
dev = [
  "pytest>=9.0.2",
  "pytest-cov>=6.0.0",
  "hypothesis>=6.0.0",
  "coverage>=7.0.0",
  "marimo>=0.11.0",
  "snoop>=0.6.0",
  "birdseye>=0.10.0",
  "hunter>=3.7.0",
  "viztracer>=0.16.0",
]

[tool.pytest.ini_options]
testpaths = [
  "tests/integration",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
package = true
